Corona Bot will answer queries related to Covid-19 pandemic.
Corona Bot will is an AI bot made using Python and implemented using Flask microframework.
It uses ChatterBot library which is is a Python library that makes it easy to generate automated responses to a user’s input. ChatterBot uses a selection of machine learning algorithms to produce different types of responses.
ChatterBot comes with a corpus data and utility module that makes it easy to quickly train your bot to communicate.