CoronaBot
CoronaBot is an AI chatbot which can be used to solve your basic queries about coronavirus.
It has been designed using Python and implemented using Flask microframework along with HTML and CSS to provide a basic look to the website.
It basically uses ChatterBot which is a Python library that makes it easy to generate automated responses to a user’s input. ChatterBot uses a selection of machine learning algorithms to produce different types of responses.